refactor: orderbook helpers
