--- packages/core/src/helpers/calcMaxVolume.ts
@@ -1,5 +1,0 @@
-import { accumulateVolume } from "./accumulateVolume";
-
-export const calcMaxVolume = (bids: string[][], asks: string[][]) => {
-  return Math.max(...accumulateVolume(bids), ...accumulateVolume(asks));
-};
